c273: api/client.ts split slice 3 — HIDSAG family (#441 2.4)
Third slice of the c261 incremental split. Extracts the 8
HIDSAG-specific types (HidsagPreprocessingSubset,
HidsagCrossPreprocessingStability, HidsagPreprocessingSensitivity,
HidsagDistribution, HidsagMethodAggregate, HidsagBlock,
HidsagMethodStatistics, HidsagEda) to api/hidsag.ts (130 LOC).

client.ts changes:
- 91 lines removed (49 from the lower HIDSAG block + 42 from the
  upper preprocessing block consolidated via re-export).
- client.ts: 1213 -> 1128 lines.
- Type re-exports preserve all existing imports.

Cumulative since c261: 1392 -> 1128 LOC (-19%).

Incidental: Benchmarks.tsx gains 2 explicit type annotations on
.map() callbacks where tsc-strict dropped inference through the
re-export chain. Two-line fix; behaviour unchanged.

Verification:
- pnpm typecheck -> clean
- pnpm test -> 30 passed
- pnpm build -> clean
